Crunch messages, handlers, queries
