--- x/domain/types/msg_delete_account.go
@@ -1,52 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MsgDeleteAccount is the request model
-// used to delete an account
-type MsgDeleteAccount struct {
-	// Domain is the name of the domain of the account
-	Domain string
-	// Name is the name of the account
-	Name string
-	// Owner is the owner of the account
-	Owner sdk.AccAddress
-}
-
-// Route implements sdk.Msg
-func (m *MsgDeleteAccount) Route() string {
-	return RouterKey
-}
-
-// Type implements sdk.Msg
-func (m *MsgDeleteAccount) Type() string {
-	return "delete_account"
-}
-
-// ValidateBasic implements sdk.Msg
-func (m *MsgDeleteAccount) ValidateBasic() error {
-	if m.Owner == nil {
-		return sdkerrors.Wrap(ErrInvalidOwner, "empty")
-	}
-	if m.Name == "" {
-		return sdkerrors.Wrap(ErrInvalidAccountName, "empty")
-	}
-	if m.Domain == "" {
-		return sdkerrors.Wrap(ErrInvalidDomainName, "empty")
-	}
-	// success
-	return nil
-}
-
-// GetSignBytes implements sdk.Msg
-func (m *MsgDeleteAccount) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners implements sdk.Msg
-func (m *MsgDeleteAccount) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{m.Owner}
-}
