--- x/domain/types/msg_delete_domain.go
@@ -1,45 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MsgDeleteDomain is the request
-// model to delete a domain
-type MsgDeleteDomain struct {
-	Domain string
-	Owner  sdk.AccAddress
-}
-
-// Route implements sdk.Msg
-func (m *MsgDeleteDomain) Route() string {
-	return RouterKey
-}
-
-// Type implements sdk.Msg
-func (m *MsgDeleteDomain) Type() string {
-	return "delete_domain"
-}
-
-// ValidateBasic implements sdk.Msg
-func (m *MsgDeleteDomain) ValidateBasic() error {
-	if m.Domain == "" {
-		return sdkerrors.Wrap(ErrInvalidDomainName, "empty")
-	}
-	if m.Owner == nil {
-		return sdkerrors.Wrap(ErrInvalidOwner, "empty")
-	}
-	// success
-	return nil
-}
-
-// GetSignBytes implements sdk.Msg
-func (m *MsgDeleteDomain) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners implements sdk.Msg
-func (m *MsgDeleteDomain) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{m.Owner}
-}
