--- x/domain/types/msg_renew_account.go
@@ -1,46 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MsgRenewAccount is the request
-// model used to renew accounts
-type MsgRenewAccount struct {
-	// Domain is the domain of the account
-	Domain string
-	// Name is the name of the account
-	Name string
-}
-
-// Route implements sdk.Msg
-func (m *MsgRenewAccount) Route() string {
-	return RouterKey
-}
-
-// Type implements sdk.Msg
-func (m *MsgRenewAccount) Type() string {
-	return "renew_account"
-}
-
-// ValidateBasic implements sdk.Msg
-func (m *MsgRenewAccount) ValidateBasic() error {
-	if m.Domain == "" {
-		return sdkerrors.Wrap(ErrInvalidDomainName, "empty")
-	}
-	if m.Name == "" {
-		return sdkerrors.Wrap(ErrInvalidAccountName, "empty")
-	}
-	return nil
-}
-
-// GetSignBytes implements sdk.Msg
-func (m *MsgRenewAccount) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners implements sdk.Msg
-func (m *MsgRenewAccount) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{}
-}
