--- x/domain/types/msg_renew_domain.go
@@ -1,41 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MsgRenewDomain is the request
-// model used to renew a domain
-type MsgRenewDomain struct {
-	// Domain is the domain name to renew
-	Domain string
-}
-
-// Route implements sdk.Msg
-func (m *MsgRenewDomain) Route() string {
-	return RouterKey
-}
-
-// Type implements sdk.Msg
-func (m *MsgRenewDomain) Type() string {
-	return "renew_domain"
-}
-
-// ValidateBasic implements sdk.Msg
-func (m *MsgRenewDomain) ValidateBasic() error {
-	if m.Domain == "" {
-		return sdkerrors.Wrapf(ErrInvalidDomainName, "empty")
-	}
-	return nil
-}
-
-// GetSignBytes implements sdk.Msg
-func (m *MsgRenewDomain) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners implements sdk.Msg
-func (m *MsgRenewDomain) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{}
-}
