--- x/domain/types/msg_transfer_account.go
@@ -1,57 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MsgTransferAccount is the request
-// model used to transfer accounts
-type MsgTransferAccount struct {
-	// Domain is the domain name of the account
-	Domain string
-	// Account is the account name
-	Name string
-	// Owner is the actual owner of the account
-	Owner sdk.AccAddress
-	// NewOwner is the new owner of the account
-	NewOwner sdk.AccAddress
-}
-
-// Route implements sdk.Msg
-func (m *MsgTransferAccount) Route() string {
-	return RouterKey
-}
-
-// Type implements sdk.Msg
-func (m *MsgTransferAccount) Type() string {
-	return "transfer_account"
-}
-
-// ValidateBasic implements sdk.Msg
-func (m *MsgTransferAccount) ValidateBasic() error {
-	if m.Domain == "" {
-		return sdkerrors.Wrap(ErrInvalidDomainName, "empty")
-	}
-	if m.Name == "" {
-		return sdkerrors.Wrap(ErrInvalidAccountName, "empty")
-	}
-	if m.Owner == nil {
-		return sdkerrors.Wrap(ErrInvalidOwner, "empty")
-	}
-	if m.NewOwner == nil {
-		return sdkerrors.Wrap(ErrInvalidOwner, "new owner is empty")
-	}
-	// success
-	return nil
-}
-
-// GetSignBytes implements sdk.Msg
-func (m *MsgTransferAccount) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners implements sdk.Msg
-func (m *MsgTransferAccount) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{m.Owner}
-}
