--- x/domain/types/msg_transfer_domain.go
@@ -1,48 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MsgTransferDomain is the request model
-// used to transfer a domain
-type MsgTransferDomain struct {
-	Domain   string
-	Owner    sdk.AccAddress
-	NewAdmin sdk.AccAddress
-}
-
-// Route implements sdk.Msg
-func (m *MsgTransferDomain) Route() string {
-	return RouterKey
-}
-
-// Type implements sdk.Msg
-func (m *MsgTransferDomain) Type() string {
-	return "transfer_domain"
-}
-
-// ValidateBasic implements sdk.Msg
-func (m *MsgTransferDomain) ValidateBasic() error {
-	if m.Domain == "" {
-		return sdkerrors.Wrap(ErrInvalidDomainName, "empty")
-	}
-	if m.Owner == nil {
-		return sdkerrors.Wrap(ErrInvalidOwner, "empty")
-	}
-	if m.NewAdmin == nil {
-		return sdkerrors.Wrap(ErrInvalidRequest, "new admin is empty")
-	}
-	return nil
-}
-
-// GetSignBytes implements sdk.Msg
-func (m *MsgTransferDomain) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners implements sdk.Msg
-func (m *MsgTransferDomain) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{m.Owner}
-}
